Drop support for legacy Python 2.7

+++ myhdl/_bin.py
@@ -20,9 +20,6 @@
 """ module with the bin function.
 
 """
-from myhdl._compat import long
-
-
 def _int2bitstring(num):
     if num == 0:
         return '0'
@@ -46,7 +43,7 @@ def bin(num, width=0):
     Optional parameter:
     width -- specifies the desired string (sign bit padding)
     """
-    num = long(num)
+    num = int(num)
     s = _int2bitstring(num)
     if width:
         pad = '0'
