+++ myhdl/_modbv.py
@@ -21,7 +21,6 @@
 from __future__ import absolute_import
 
 from ._intbv import intbv
-from ._compat import long
 
 
 class modbv(intbv):
@@ -54,7 +53,7 @@ class modbv(intbv):
             if i <= j:
                 raise ValueError("modbv[i:j] requires i > j\n"
                                  "            i, j == %s, %s" % (i, j))
-            res = modbv((self._val & (long(1) << i) - 1) >> j, _nrbits=i - j)
+            res = modbv((self._val & (1 << i) - 1) >> j, _nrbits=i - j)
             return res
         else:
             i = int(key)
